Making question class
Question will be used for most things related to the question generation like getting/formatting questions & generating questions

This added the functionality of parsing a txt document to get all the info for an equation question.

Altered the questions I was using to have a cleaner and tighter format.

Various other small fixes or deletes